requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & OAuth
authlib==1.4.0
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Background Tasks
celery[redis]==5.4.0
redis==5.2.0

# Security & Encryption
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email
postmarkhttpclient==0.5.5

# AI
openai==1.54.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code Quality
bandit==1.8.0
mypy==1.13.0

# Monitoring
sentry-sdk[fastapi]==2.19.2
